refactor(AdjustmentsPanel): split inline node controls into per-type files with dispatch map (#463)
Extract 13 inline adjustment control components from AdjustmentsPanel.tsx
(648 lines) into individual files under controls/. Replace the 14-case
switch statement in NodeControls with a NODE_CONTROLS_MAP lookup.
Main file drops to 282 lines.
