+++ src/panels/AdjustmentsPanel/controls/types.ts
@@ -1,0 +1,6 @@
+import type { AdjustmentNode } from '../../../types/adjustment-nodes';
+
+export interface NodeControlProps<T extends AdjustmentNode = AdjustmentNode> {
+  node: T;
+  onChange: (params: Partial<AdjustmentNode>) => void;
+}
